flytter på ting og legger til user middleware
